Make common variables list-like
Declare CFLAGS, CPPFLAGS, and LDFLAGS as list variables for convenience
in build and module configurations. Also set the export flag on them so
that they are available to make from the environment. They can still be
overridden for a particular build from the command line, like a typical
make invocation.