Auto-generate rules for creating directories
Extend the post-processing awk script to also identify targets that are
located somewhere in the target directory or which explicitly depend on
a directory in the target. All targets should be prefixed with the make
expanded variable TARGET. By expanding the TARGET variable in make
instead of in shell the choice is delayed and it's possible to do some
configuration of the build even with only the makefile. Similarly, all
target sources should have either $(TARGET) or $(BUILD_HOME) as a
prefix. For instance, rules should look like

    $(TARGET)/obj/foo/bar.o: $(BUILD_HOME)/src/foo/bar.c ...

More liberal uses of target directories is now finally possible.