Remove BUILD_SOURCE_DIR
Make the MODULES variable contain module directory paths rather than
module names. The paths must be relative to the build configuration
file. The basename(1) of the path is treated as the name of the module,
but it's now possible to place modules in more than one directory (the
target directory has a flat namespace, so module names still need to be
unique).

Other changes
- Rename BUILD_HOME into SOURCE in both shell and make
- BUILD_INCLUDE_DIR is now SOURCE_INCLUDE, and it can be zero or more
  directories in the source directory which contain headers
- Removed DEPEND shell variable
- Remove MODULE_{SOURCE,OBJECT}_DIR shell variables now that the use of
  directories is easier. The need for creating variables for this will
  become more clear later.