List object files and recreate archive every time
Introduce the OBJECT.$MODULE make variable which contains a list of
every object in the module. Base the lib$MODULE.a rules on this and
recreate the archive from scratch every time rather than incrementally
upating it with the ? make variable. This is a partial solution for
removing objects from archives when the source file has been deleted.