Add regex to make matching targets
Parse command line for arguments of the form "make:<regex>" and
"show:<option>". Targets with a colon in them are not widely supported
by make implementations and should be safe to use for this purpose and
for pseudo targets. Change the "Makefile" target into "show:Makefile".
Also add "show:targets" which lists all the defined make targets on
standard output.